Refactored methods generating training sets.
Basically, I figured out what was the reason of non-perfect weights
vectors. It was due to non-regular data distribution caused by
pseudo-random data generating process. Instead of generating milions of
examples, now there are only 4 (or two in case of single Perceptron)
samples generated but they are regularly distributed and this makes the
weights vectors place perfectly on X and Y axis.